+++ packages/tiptap-extensions/web/src/paragraph.ts
@@ -66,14 +66,7 @@ export const paragraph: RicosExtension = {
         return [
           'div',
           attrs,
-          [
-            'span',
-            {
-              class: styles['text-wrapper'],
-              ...(shouldAddDir ? { dir: 'auto' } : {}),
-            },
-            0,
-          ],
+          ['span', { ...(shouldAddDir ? { dir: 'auto' } : {}) }, 0],
         ] as DOMOutputSpec;
       },